Standardise on MinervaNeue's skin name (minerva)
We will standardise on the key "minerva" and the styles
in MediaWiki:Minerva.css for desktop since this is what is
used for mobile. The name will however appear as MinervaNeue
to be consistent with the skin repo name (not the code).

Please merge I5fdb21243e0b599b07aa228107fd4b8c306c696c
shortly after merging this to remove the remaining logic
relating to this from MobileFrontend as otherwise two skins
will appear in user preferences.

Any existing users of MinervaNeue (key minervaneue on desktop)
will be thrown back into the Vector skin and will need to opt
into the MinervaNeue skin again. This is a minor disruption
given the low amount of users and instability of the skin.

Bug: T171644
Change-Id: Ieffb6334168047dcf05dcb60281865ab47056dc0