QA: Stabilise browser tests
Changes:
* Default skin is not configured so noticed on test runs it's using
Default skin for desktop mode. Given this has no footer and way to
switch to mobile this could be problematic in future.
* Remove empty README
* Drop suggested language integration test - setting up interwiki links currently
happens via InterwikiLoadPrefix hook however this seems to be unreliable. The beta
cluster is a much more reliable place to test language links so let's rely on that
instead.
* Drop unnecessary heading check from "Successful edit" scenario. It's unnecessary
and flakey (presumably given the title doesn't change)
* Rename test scenario and remove a duplicate scenario

Depends-On: I888b3c546f77fa350853a7bf9bfbfbeb8ed6de67
Change-Id: I45792a95df7fd4c3299accbffadfa447baefe0ce