Doc: move config type and defaults near identity
The readme describes all of Minerva's configs and includes a type and
what the default value is at the end of each section. Since the
descriptions for some configs is lengthy, moving the type and default
next to each config's name gives a consistent location for the
information and allows for quick look up without parsing.

Change-Id: I1583c11c80d9b775d93f59302eb2666b754fb412