Set some authors inside Minerva
Apologies in advance if any unintentional hurt feelings!
This is based on a git blame of the skins directory in
I985c4e3a88b59461d471945ccf74cd291db45a61 before the Minerva move
and consulting the extension.json of MobileFrontend.

I've listed anyone with over 100 commits to the MobileFrontend resources/
skins folder as an author.

The Readers Web Team as the current maintainers are listed first.

If you feel you should be listed as an author of this skin
as someone who has contributed more than 100 commits to the skin code
(not MobileFrontend code), apologies and please add yourself.

Change-Id: Ifffb07d1decb47a5e08b9c45c88ccf106f0797a1