QA: Run references test per commit
Follow up to Id312638d86179e75bc670e72e5943f8c00232bbb
which switched the #content area from a div to a main
HTML element causing this to fail to match.

Test suite will now also run on @integration so that
this kind of issue is not caught too late in the beta
cluster.

Additional changes:
* Merge two similar tests

Bug: T201956
Change-Id: I1d402aaebc40dcca61979aa521cd8e1a1ce274d9